fix roll civs

+++ internal/draft.go
@@ -16,6 +16,7 @@ var BannedLeaders = []string{
 	"HAMMURABI",
 }
 
+// TODO: apply min-tiers rule
 func OfferPicks(db *DatabaseOperations, draft domain.Ci6ndexDraft, numPlayers int) ([]DraftOffering, error) {
 	ds, err := db.queries.GetDraftStrategy(context.Background(), draft.DraftStrategy)
 	if err != nil {
@@ -63,7 +64,7 @@ func OfferPicks(db *DatabaseOperations, draft domain.Ci6ndexDraft, numPlayers in
 		if rules["pool_size"] == nil {
 			psize = 1
 		} else {
-			psize = rules["pool_size"].(int)
+			psize = int(rules["pool_size"].(float64))
 		}
 
 		i := 1
